Add method for virtual cam and refactor code
get_next_frame() is the default method to get frames from the pi stream.
It blocks until there is a new frame available from the stream. So it naturally syncs with the fps of the physical cam.

If you want to manually controll the rate at wich the frames get read from the stream you can use the method get_next_frame_at_desired_rate(self).
This adjusts the rate at wich new frames get read from the stream to the desired framerate.
This gets used when virtual cam is selected. So the pre-recorded video gets read at the correct speed. Otherwise it would read the frames as fast as possible.

Also cleaned up the code a bit and wrote comments.