Add timestamping to frames
Camera Class now creates a Timestamp for each frame as soon as it recieves the frame from the cam stream.
This could propably be done by the physical camera for a more precise match between the timestamp and the frame.
But should be good enough if its done here.

Also refactored the code a bit